finally

+++ Models/ProcessResponse.cs
@@ -10,10 +10,11 @@ namespace ReturnOrderPortal.Models
     {
         public int RequestId { get; set; }
 
-        public double ProcessingCharge { get; set; }
+        public int ProcessingCharge { get; set; }
 
-        public double PackagingAndDeliveryCharge { get; set; }
+        public int PackagingAndDeliveryCharge { get; set; }
 
         public DateTime DateOfDelivery { get; set; }
+
     }
 }
